+++ src/searchAndSort/interfaces.ts
